UNBReader successfully read and create a list of all courses.
Next step will be to make it write that list (which is a schedule object
full of all the courses) to a file. The file will not be a .cvs file, it
will instead just contain a serializable Schedule object holding all the
courses. This should make it easier to read it back in later.