Fixed bug in AddEditClassTime: able to add a conflicting class time.
Bug Description:
When adding a class time to a section, you were able to add a class time
that conflicts with another class time already in this section.
Ex: Adding
	Lab MWF 10:00-12:00
when section already contains
	Lecture MWF 10:30-12:30
(Class times in the same section should never overlap!)

Bug Fix:
Added a check to make sure the class time you are trying to add does not
conflict with any already present class time in this section.