Implemented a second algorithm, refactored a lot, added basic tests.
Note: In theory, the second version of the algorithm is not always better. But, for "normal" cases, it is usually much better than the first one. Because of that, I have it currently set up so it always uses algorithm V2 by default, but you can always force it to use algorithm 1 by using the alternate method in the class.